requirements.txt
cffi==1.16.0
evdev==1.7.0
pycparser==2.22
pynput==1.7.6
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
PySide6==6.6.3.1
PySide6_Addons==6.6.3.1
PySide6_Essentials==6.6.3.1
python-xlib==0.33
shiboken6==6.6.3.1
six==1.16.0
system-hotkey==1.0.3
xcffib==1.5.0
xpybutil==0.0.6
google-generativeai==0.8.3
python-dotenv==1.0.0
pydantic==2.9.2
